Refactor: Split redemption service into smaller files
Split the redemption service into smaller, more manageable files for improved code organization and maintainability.
